requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
